Refactored nodes into individual files for better code clarity and refactoring.
