+++ nodes/DiffusersSampler.py
@@ -1,0 +1,68 @@
+import torch
+from comfy.model_management import get_torch_device
+from ..utils import convert_images_to_tensors
+
+class DiffusersSampler:
+    def __init__(self):
+        self.torch_device = get_torch_device()
+
+    @classmethod
+    def INPUT_TYPES(s):
+        return {
+            "required": {
+                "pipeline": ("PIPELINE",),
+                "prompt": (
+                    "STRING",
+                    {"multiline": True, "default": "a photo of a cat"},
+                ),
+                "negative_prompt": (
+                    "STRING",
+                    {
+                        "multiline": True,
+                        "default": "watermark, ugly, deformed, noisy, blurry, low contrast",
+                    },
+                ),
+                "width": ("INT", {"default": 768, "min": 1, "max": 2048, "step": 1}),
+                "height": ("INT", {"default": 768, "min": 1, "max": 2048, "step": 1}),
+                "steps": ("INT", {"default": 50, "min": 1, "max": 2000}),
+                "cfg": (
+                    "FLOAT",
+                    {
+                        "default": 7.0,
+                        "min": 0.0,
+                        "max": 100.0,
+                        "step": 0.1,
+                        "round": 0.01,
+                    },
+                ),
+                "seed": ("INT", {"default": 0, "min": 0, "max": 0xFFFFFFFFFFFFFFFF}),
+            }
+        }
+
+    RETURN_TYPES = ("IMAGE",)
+
+    FUNCTION = "sample"
+
+    CATEGORY = "Diffusers"
+
+    def sample(
+        self,
+        pipeline,
+        prompt,
+        negative_prompt,
+        height,
+        width,
+        steps,
+        cfg,
+        seed,
+    ):
+        images = pipeline(
+            prompt=prompt,
+            height=height,
+            width=width,
+            num_inference_steps=steps,
+            guidance_scale=cfg,
+            negative_prompt=negative_prompt,
+            generator=torch.Generator(self.torch_device).manual_seed(seed),
+        ).images
+        return (convert_images_to_tensors(images),)
